change return type - double to BigDecimal

+++ src/main/java/domain/amount/FixedDepositAmount.java
@@ -1,5 +1,7 @@
 package domain.amount;
 
+import java.math.BigDecimal;
+
 import domain.interest_rate.InterestRate;
 
 public class FixedDepositAmount implements LumpSumInvestmentAmount {
@@ -29,7 +31,7 @@ public class FixedDepositAmount implements LumpSumInvestmentAmount {
 	}
 
 	@Override
-	public double calMonthlyInterest(InterestRate interestRate) {
-		return interestRate.calMonthlyInterest(amount).doubleValue();
+	public BigDecimal calMonthlyInterest(InterestRate interestRate) {
+		return interestRate.calMonthlyInterest(amount);
 	}
 }
